another try with a wizard step component
