+++ modules/gui/src/de/diedavids/cuba/wizard/gui/components/WizardStep.java
@@ -1,0 +1,7 @@
+package de.diedavids.cuba.wizard.gui.components;
+
+import com.haulmont.cuba.gui.components.VBoxLayout;
+
+public interface WizardStep extends VBoxLayout {
+    String NAME = "step";
+}
