+++ modules/web/src/de/diedavids/cuba/wizard/web/gui/components/WebWizardStep.java
@@ -1,0 +1,9 @@
+package de.diedavids.cuba.wizard.web.gui.components;
+
+import com.haulmont.cuba.web.gui.components.WebVBoxLayout;
+import de.diedavids.cuba.wizard.gui.components.WizardStep;
+
+public class WebWizardStep extends WebVBoxLayout implements WizardStep {
+    public WebWizardStep() {
+    }
+}
